Use isolated environment variables in docker-compose:
Both pg-pool and node-postgres are compatible with certain
environment variables similar to those supported by 'psql'.
However, this compatibility is not guaranteed, and there is
a risk of these variables being inadvertently modified in an
interactive shell when a user launches an API server from the
command line. To ensure consistency, the environment variables
have been synchronized with those used in 'kgi-processing' and
incorporated into the Docker Compose environment settings.